+++ library/src/main/java/me/denley/preferencebinder/internal/Binding.java
@@ -1,4 +1,4 @@
-package me.denley.preferenceinjector.internal;
+package me.denley.preferencebinder.internal;
 
 import java.lang.annotation.ElementType;